LayoutTests/svg/animations/animate-color-fill-from-by-expected.txt

SVG 1.1 dynamic animation tests

Tests animation on 'currentColor'.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".



Initial condition:
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 208
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS fillPaint.uri is ""
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR

Half-time condition:
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 208
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 104
PASS fillPaint.uri is ""
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR

End condition:
PASS color.red.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 208
PASS color.green.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 0
PASS color.blue.getFloatValue(CSSPrimitiveValue.CSS_NUMBER) is 208
PASS fillPaint.uri is ""
PASS fillPaint.paintType is SVGPaint.SVG_PAINTTYPE_CURRENTCOLOR
PASS fillPaint.colorType is SVGColor.SVG_COLORTYPE_CURRENTCOLOR
PASS successfullyParsed is true

TEST COMPLETE
